Bug 1973417 - Part 3. Allow draining for FFmpegVideoDecoder to be split across multiple calls.
On Android, we have a limited output buffer pool which can easily be
exhausted while draining the decoder. The existing plumbing allows for
partial drains, and the caller is expected to call Drain until it
returns no samples. FFmpegVideoDecoder always early returns after the
first drain due to the EOS status when we attempt to queue the EOS empty
packet.

This patch changes it so that we only supply the EOS empty packet once,
and instead rely upon the receive frame plumbing to either return more
frames (because our last drain was incomplete due to buffering limits)
or return nothing (because the decoder has truly been drained).

Differential Revision: https://phabricator.services.mozilla.com/D254628