Bug 1980282 - Batch gfxVar updates during fallback and hardware video changes.
When we update gfxVars, if we create gfxVarsCollectUpdates on the stack,
we can consolidate those updates into a single notification and IPC
message per process. This will save on churn, particularly for the media
processes, which may want to monitor multiple gfxVars for changes
affecting hardware video decoding/encoding.

This also moves the CodecSupportInfo from a gfxVar into a member of
GfxInfoBase since that is the only place that needs access to it. This
avoids bouncing back a gfxVar update back to all of the child processes
after a codec support update from each of them.

Differential Revision: https://phabricator.services.mozilla.com/D259335