Bug 1980802 - Ensure we cleanup AVCodecContext correctly in FFmpegDataDecoder/Encoder.
On older ffmpeg versions, we need to call avcodec_close and av_freep
with mCodecContext. We also need to explicitly free any extradata for
the decoders because it avcodec_close only does that for the encoders.
On new ffmpeg versions, we can just call avcodec_free_context to do
everything necessary.

Differential Revision: https://phabricator.services.mozilla.com/D259815