Bug 1984668 - Add script to extract list of Windows system calls from GMP plugins.
This script should help us quickly determine whether or not there are
additional system calls for Widevine plugins when we receive an update
from Google. Since the updates are a black box, and our sandboxing might
restrict relevant APIs, or the APIs might not be available on Windows
7/8.1 for ESR 115, it is handy to be able to generate a list of methods
that are linked against, or gotten via GetProcAddress.

The output of the script is just a list of the Windows system method
names, alphabetically ordered. This allows one to easily run the script
on the current Widevine plugin, the latest we want to upgrade to and
perform a diff between them to see what has changed.

Differential Revision: https://phabricator.services.mozilla.com/D262152