Bug 1900626 - Allow Android to use our bundled clearkey CDM.
This patch makes it so that we include libclearkey.so with our APK and
so that our existing GMP plumbing can launch using that library on
Android. On other platforms we have a specific plugin file system layout
that we follow and unfortunately that is not easy to do on Android. As
such we manually setup the plugin and force the GMP plugin process to
find/load the library from our environment instead of using the absolute
path.

Differential Revision: https://phabricator.services.mozilla.com/D270967