Bug 1997848 - Fallback from acceleration for compositor session creation errors.
Simulating device resets for parent process compositing is problematic
because we need to dispatch to handle device resets, but the compositor
session creation never lets go of the main thread to allow those events
to be processed.

In reality most platforms have a GPU process now, and if we disable the
GPU process, we use Software WebRender, for which a simulated device
reset is unlikely to help anyways. The only exception is on Linux, but
it is not that common to fail to create a compositor, so it is probably
okay to fall back to Software WebRender right away.

Similarly, killing the GPU process is also a problem for the same
reason. We need to clean up inline, and there are some reentrancy
concerns. Once we investigate those, we can land a later patch to
re-enable this for the compositor session path.

Differential Revision: https://phabricator.services.mozilla.com/D271214